+++ src/lib.rs
@@ -36,13 +36,14 @@ pub use crate::metadata::{Metadata21, WheelMetadata};
 pub use crate::module_writer::{
     write_dist_info, ModuleWriter, PathWriter, SDistWriter, WheelWriter,
 };
+pub use crate::pyproject_toml::PyProjectToml;
 pub use crate::python_interpreter::PythonInterpreter;
 pub use crate::read_distribution::{
     get_metadata_for_distribution, get_supported_version_for_distribution,
 };
 pub use crate::target::Target;
 pub use auditwheel::Manylinux;
-pub use source_distribution::{get_pyproject_toml, source_distribution};
+pub use source_distribution::source_distribution;
 #[cfg(feature = "upload")]
 pub use {
     crate::registry::Registry,
@@ -57,6 +58,7 @@ mod compile;
 mod develop;
 mod metadata;
 mod module_writer;
+mod pyproject_toml;
 mod python_interpreter;
 mod read_distribution;
 #[cfg(feature = "upload")]
